Double to float conversion of the code
git-svn-id: http://192.168.1.75/svn/gpair@12 48422db7-73d2-4338-a790-e04df5fead5b

+++ gpair.h
@@ -4,5 +4,5 @@
 int read_oifits( );
 void vis2data( );
 void image2vis( );
-double data2chi2( );
-void write_fits_image( double* image, int* status );
+float data2chi2( );
+void write_fits_image( float* image, int* status );
